Add files via upload
In this project we are designing a multimeter circuit and the value of the measured reading will be directly displayed on the spectacles wore by the user with the help of transmitting Bluetooth module. In this project we are trying to safeguard and make the life simpler for the workers, who work on voltage systems with the help of smart technology.